test(e2e): grant the OIDC identity admin rights on the remote cluster
On-behalf-of: @SAP <mikolaj.kucinski@sap.com>
